Cash Register Scene
Still Has some problems. It does return you to the menus the first time,
after that it throws errors. It does wait a little bit before doing so
that we can see the score total up. The camera is still having some
trouble being moved to and looking at the right locations.